Refactor UI to separate files
