+++ src/lib/components/ASSSubtitleProcessor.svelte
@@ -1,66 +1,11 @@
 <script lang="ts">
-import { perma_state } from "$lib/persistent-storage.svelte"
 import { temp_state } from "$lib/temporary-storage.svelte"
 // See http://www.tcax.org/docs/ass-specs.htm
-import {
-    available_fonts,
-    download_video,
-    format_time_remaining,
-    generate_ass_file,
-    process_ass_subtitles,
-    render_ass_frame_preview,
-} from "$lib/utils/subtitle-processing"
-import FontPreview from "./FontPreview.svelte"
-
-let sample_text = $state("The quick brown fox jumps over the lazy dog")
-
-function handle_video_upload(event: Event) {
-    const target = event.target as HTMLInputElement
-    if (target.files?.[0]) {
-        temp_state.ffmpeg.video_file = target.files[0]
-        if (temp_state.ffmpeg.preview_url) {
-            URL.revokeObjectURL(temp_state.ffmpeg.preview_url)
-            temp_state.ffmpeg.preview_url = null
-        }
-    }
-}
-
-function handle_srt_upload(event: Event) {
-    const target = event.target as HTMLInputElement
-    if (target.files?.[0]) {
-        temp_state.ffmpeg.srt_file = target.files[0]
-        if (temp_state.ffmpeg.preview_url) {
-            URL.revokeObjectURL(temp_state.ffmpeg.preview_url)
-            temp_state.ffmpeg.preview_url = null
-        }
-    }
-}
-
-async function download_ass_file_wrapper() {
-    if (!temp_state.ffmpeg.srt_file) {
-        temp_state.ffmpeg.error_message = "Please upload an SRT file first"
-        return
-    }
-
-    try {
-        const srt_content = await temp_state.ffmpeg.srt_file.text()
-        const ass_content = generate_ass_file(srt_content)
-
-        const blob = new Blob([ass_content], { type: "text/plain" })
-        const url = URL.createObjectURL(blob)
-        const a = document.createElement("a")
-        a.href = url
-        a.download = `styled-subtitles-${temp_state.ffmpeg.video_file?.name.replace(/\.[^/.]+$/, "") || "video"}.ass`
-        document.body.appendChild(a)
-        a.click()
-        document.body.removeChild(a)
-        URL.revokeObjectURL(url)
-
-        temp_state.ffmpeg.message = "ASS file downloaded successfully"
-    } catch (error) {
-        temp_state.ffmpeg.error_message = `Failed to download ASS file: ${error}`
-    }
-}
+import { format_time_remaining, process_ass_subtitles } from "$lib/utils/subtitle-processing"
+import SubtitleFramePreview from "./ui/SubtitleFramePreview.svelte"
+import SubtitleOutput from "./ui/SubtitleOutput.svelte"
+import SubtitleSettings from "./ui/SubtitleSettings.svelte"
+import SubtitleUpload from "./ui/SubtitleUpload.svelte"
 </script>
 
 <div class="flex flex-col space-y-2">
@@ -69,345 +14,9 @@ async function download_ass_file_wrapper() {
     <div
         class="processor-container mx-auto max-w-6xl min-w-0 space-y-6 rounded-xl bg-white p-4 shadow-lg flex flex-col"
     >
-        <!-- Upload Section -->
-        <div class="rounded-lg border border-green-200 bg-gradient-to-r from-green-50 to-emerald-50 p-4">
-            <h2 class="mb-4 text-lg font-bold text-gray-800">Upload Files</h2>
-
-            <!-- Video Upload -->
-            <div class="mb-4">
-                <label
-                    for="video-upload"
-                    class="mb-2 block text-sm font-semibold text-gray-700"
-                    >Video</label
-                >
-                <div class="relative">
-                    <input
-                        id="video-upload"
-                        type="file"
-                        accept="video/*"
-                        onchange={handle_video_upload}
-                        class="w-full rounded-lg border-2 border-dashed border-gray-300 px-3 py-4 transition-all duration-200 file:mr-2 file:rounded file:border-0 file:bg-green-50 file:px-3 file:py-1 file:text-xs file:font-semibold file:text-green-700 hover:border-green-400 focus:ring-2 focus:ring-green-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-                {#if temp_state.ffmpeg.video_file}
-                    <p
-                        class="mt-1 inline-flex items-center rounded bg-green-100 px-2 py-1 text-xs font-medium text-green-700"
-                    >
-                        {temp_state.ffmpeg.video_file.name}
-                    </p>
-                {/if}
-            </div>
-
-            <!-- SRT Upload -->
-            <div class="mb-4">
-                <label
-                    for="srt-upload"
-                    class="mb-2 block text-sm font-semibold text-gray-700"
-                    >SRT Subtitles</label
-                >
-                <div class="relative">
-                    <input
-                        id="srt-upload"
-                        type="file"
-                        accept=".srt"
-                        onchange={handle_srt_upload}
-                        class="w-full rounded-lg border-2 border-dashed border-gray-300 px-3 py-4 transition-all duration-200 file:mr-2 file:rounded file:border-0 file:bg-green-50 file:px-3 file:py-1 file:text-xs file:font-semibold file:text-green-700 hover:border-green-400 focus:ring-2 focus:ring-green-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-                {#if temp_state.ffmpeg.srt_file}
-                    <p
-                        class="mt-1 inline-flex items-center rounded bg-green-100 px-2 py-1 text-xs font-medium text-green-700"
-                    >
-                        {temp_state.ffmpeg.srt_file.name}
-                    </p>
-                {/if}
-            </div>
-        </div>
-
-        <!-- Settings Section -->
-        <div class="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
-            <!-- Basic Font Settings -->
-            <div class="rounded-lg border border-blue-200 bg-gradient-to-r from-blue-50 to-indigo-50 p-6">
-                <h2 class="mb-4 text-xl font-bold text-gray-800">Font Settings</h2>
-
-                <!-- Font Selection -->
-                <div class="mb-4">
-                    <label
-                        for="font-select"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Font Family</label
-                    >
-                    <select
-                        id="font-select"
-                        bind:value={perma_state.subtitle_settings.font.index}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-blue-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                        {#each available_fonts as font, index}
-                            <option value={index}>{font.select_name}</option>
-                        {/each}
-                    </select>
-                </div>
-
-                <!-- Font Size -->
-                <div class="mb-4">
-                    <label
-                        for="font-size"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Font Size: {perma_state.subtitle_settings.font.size}px</label
-                    >
-                    <input
-                        id="font-size"
-                        type="range"
-                        min="12"
-                        max="120"
-                        bind:value={perma_state.subtitle_settings.font.size}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-blue-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-
-                <!-- RTL Support -->
-                <div class="mb-4">
-                    <label class="flex items-center space-x-3">
-                        <input
-                            type="checkbox"
-                            bind:checked={perma_state.subtitle_settings.font.right_to_left}
-                            class="h-5 w-5 rounded border-gray-300 text-blue-600 focus:ring-blue-500"
-                            disabled={temp_state.ffmpeg.is_processing}
-                        >
-                        <span class="text-sm font-semibold text-gray-700">Enable Right-to-Left (RTL) Text</span>
-                    </label>
-                </div>
-            </div>
-
-            <!-- Text Color Settings -->
-            <div class="rounded-lg border border-purple-200 bg-gradient-to-r from-purple-50 to-pink-50 p-6">
-                <h2 class="mb-4 text-xl font-bold text-gray-800">Text Colors</h2>
-
-                <!-- Text Color -->
-                <div class="mb-4">
-                    <label
-                        for="text-color"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Text Color</label
-                    >
-                    <div class="flex items-center space-x-3">
-                        <input
-                            id="text-color"
-                            type="color"
-                            bind:value={perma_state.subtitle_settings.text.color}
-                            class="h-12 w-20 rounded-lg border border-gray-300 cursor-pointer"
-                            disabled={temp_state.ffmpeg.is_processing}
-                        >
-                        <span class="text-sm text-gray-600">{perma_state.subtitle_settings.text.color}</span>
-                    </div>
-                </div>
-
-                <!-- Stroke Color -->
-                <div class="mb-4">
-                    <label
-                        for="stroke-color"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Stroke Color</label
-                    >
-                    <div class="flex items-center space-x-3">
-                        <input
-                            id="stroke-color"
-                            type="color"
-                            bind:value={perma_state.subtitle_settings.text.stroke}
-                            class="h-12 w-20 rounded-lg border border-gray-300 cursor-pointer"
-                            disabled={temp_state.ffmpeg.is_processing}
-                        >
-                        <span class="text-sm text-gray-600">{perma_state.subtitle_settings.text.stroke}</span>
-                    </div>
-                </div>
-
-                <!-- Stroke Size -->
-                <div class="mb-4">
-                    <label
-                        for="stroke-size"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Outline Size: {perma_state.subtitle_settings.text.outline_size}px</label
-                    >
-                    <input
-                        id="stroke-size"
-                        type="range"
-                        min="0"
-                        max="10"
-                        bind:value={perma_state.subtitle_settings.text.outline_size}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-purple-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-            </div>
-
-            <!-- Shadow & Position Settings -->
-            <div class="rounded-lg border border-emerald-200 bg-gradient-to-r from-emerald-50 to-teal-50 p-6">
-                <h2 class="mb-4 text-xl font-bold text-gray-800">Shadow & Position</h2>
-
-                <!-- Shadow Blur -->
-                <div class="mb-4">
-                    <label
-                        for="shadow-blur"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Shadow: {perma_state.subtitle_settings.shadow.size}px</label
-                    >
-                    <input
-                        id="shadow-blur"
-                        type="range"
-                        min="0"
-                        max="4"
-                        bind:value={perma_state.subtitle_settings.shadow.size}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-
-                <!-- Horizontal Margin -->
-                <div class="mb-4">
-                    <label
-                        for="position-y"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Horizontal Margin: {perma_state.subtitle_settings.position.horizontal_margin}px</label
-                    >
-                    <input
-                        id="margin-h"
-                        type="range"
-                        min="0"
-                        max="1080"
-                        bind:value={perma_state.subtitle_settings.position.horizontal_margin}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                </div>
-
-                <!-- Vertical Anchor -->
-                <div class="mb-4">
-                    <label
-                        for="anchor-y"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Vertical Anchor</label
-                    >
-                    <select
-                        id="anchor-y"
-                        bind:value={perma_state.subtitle_settings.position.vertical_anchor}
-                        class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-blue-500 focus:outline-none"
-                        disabled={temp_state.ffmpeg.is_processing}
-                    >
-                        {#each ["Top", "Center", "Bottom"] as anchor_option}
-                            <option value={anchor_option.toLowerCase()}>{anchor_option}</option>
-                        {/each}
-                    </select>
-                </div>
-
-                <!-- Position Y -->
-                {#if perma_state.subtitle_settings.position.vertical_anchor !== "center"}
-                    <div class="mb-4">
-                        <label
-                            for="position-y"
-                            class="mb-2 block text-sm font-semibold text-gray-700"
-                            >Vertical Position: {perma_state.subtitle_settings.position.vertical}px</label
-                        >
-                        <input
-                            id="position-y"
-                            type="range"
-                            min="0"
-                            max="1080"
-                            bind:value={perma_state.subtitle_settings.position.vertical}
-                            class="w-full rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                            disabled={temp_state.ffmpeg.is_processing}
-                        >
-                    </div>
-                {/if}
-            </div>
-        </div>
-
-        <!-- Font Preview -->
-        <div class="rounded-lg border border-orange-200 bg-gradient-to-r from-orange-50 to-amber-50 p-6">
-            <h3 class="mb-4 text-lg font-bold text-gray-800">Font Preview</h3>
-            <div class="rounded-lg bg-white p-2 shadow-inner">
-                <div class="mx-auto flex w-full items-center justify-center space-x-4">
-                    <FontPreview
-                        font_name={available_fonts[perma_state.subtitle_settings.font.index]
-						.font_family}
-                        font_size={perma_state.subtitle_settings.font.size}
-                        {sample_text}
-                        font_weight={available_fonts[perma_state.subtitle_settings.font.index]
-						.font_weight}
-                    />
-                </div>
-            </div>
-        </div>
-
-        <!-- Frame Preview Section -->
-        {#if temp_state.ffmpeg.video_file && temp_state.ffmpeg.srt_file}
-            <div class="rounded-lg border border-indigo-200 bg-gradient-to-r from-indigo-50 to-blue-50 p-6">
-                <h2 class="mb-4 text-xl font-bold text-gray-800">Frame Preview</h2>
-                <!-- Preview Timestamp Input -->
-                <div class="mb-6">
-                    <label
-                        for="preview-timestamp"
-                        class="mb-2 block text-sm font-semibold text-gray-700"
-                        >Preview Timestamp</label
-                    >
-                    <div class="flex items-center space-x-3">
-                        <input
-                            id="preview-timestamp"
-                            type="text"
-                            bind:value={temp_state.ffmpeg.preview_timestamp}
-                            placeholder="00:00:00"
-                            class="flex-1 rounded-lg border border-gray-300 bg-white px-4 py-3 shadow-sm transition-all duration-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                            disabled={temp_state.ffmpeg.is_processing ||
-							temp_state.ffmpeg.is_rendering_preview}
-                        >
-                        <button
-                            onclick={render_ass_frame_preview}
-                            disabled={temp_state.ffmpeg.is_processing ||
-							temp_state.ffmpeg.is_rendering_preview}
-                            class="transform rounded-lg bg-gradient-to-r from-blue-600 to-blue-700 px-6 py-3 font-semibold text-white shadow-md transition-all duration-200 hover:-translate-y-0.5 hover:from-blue-700 hover:to-blue-800 hover:shadow-lg disabled:cursor-not-allowed disabled:from-gray-400 disabled:to-gray-500"
-                        >
-                            {#if temp_state.ffmpeg.is_rendering_preview}
-                                <span class="flex items-center justify-center">
-                                    Rendering...
-                                    <div
-                                        class="ml-2 h-5 w-5 animate-spin rounded-full border-2 border-white border-t-transparent"
-                                    ></div>
-                                </span>
-                            {:else}
-                                Render Preview
-                            {/if}
-                        </button>
-                    </div>
-                </div>
-                {#if temp_state.ffmpeg.preview_url}
-                    <div class="text-center">
-                        <div class="mx-auto max-w-4xl rounded-lg bg-black p-1">
-                            <img
-                                src={temp_state.ffmpeg.preview_url}
-                                alt="ASS frame preview with subtitles"
-                                class="w-full rounded-lg"
-                                style="max-height: 500px;"
-                            >
-                        </div>
-                        <p
-                            class="mt-3 inline-block rounded-full bg-gray-100 px-4 py-2 text-sm font-medium text-gray-700"
-                        >
-                            Preview at {temp_state.ffmpeg.preview_timestamp}
-                        </p>
-                    </div>
-                {:else}
-                    <div class="rounded-lg bg-gray-50 py-8 text-center">
-                        <p class="mb-2 text-lg text-gray-500">No preview yet</p>
-                        <p class="text-sm text-gray-400">
-                            Click "Render Preview" to see how your styled subtitles will look on the video frame.
-                        </p>
-                    </div>
-                {/if}
-            </div>
-        {/if}
+        <SubtitleUpload />
+        <SubtitleSettings />
+        <SubtitleFramePreview />
 
         <!-- Processing Section -->
         <div class="rounded-lg border border-green-200 bg-gradient-to-r from-green-50 to-emerald-50 p-6">
@@ -506,65 +115,7 @@ async function download_ass_file_wrapper() {
             </button>
         </div>
 
-        <!-- Output Section -->
-        {#if temp_state.ffmpeg.output_url}
-            <div
-                class="space-y-6 rounded-lg border border-emerald-200 bg-gradient-to-r from-emerald-50 to-green-50 p-6"
-            >
-                <h2 class="text-2xl font-bold text-gray-800">ASS Output Ready!</h2>
-                <div class="overflow-hidden rounded-xl bg-black shadow-2xl">
-                    <video
-                        src={temp_state.ffmpeg.output_url}
-                        controls
-                        class="mx-auto w-full max-w-4xl"
-                    >
-                        Your browser does not support the video tag.
-                    </video>
-                </div>
-                <div class="flex flex-col gap-4 border-t border-emerald-200 pt-4 sm:flex-row">
-                    <button
-                        onclick={download_ass_file_wrapper}
-                        class="flex flex-1 transform items-center justify-center rounded-lg bg-gradient-to-r from-purple-600 to-purple-700 px-8 py-4 text-lg font-bold text-white shadow-lg transition-all duration-200 hover:-translate-y-1 hover:from-purple-700 hover:to-purple-800 hover:shadow-xl"
-                    >
-                        Download ASS File
-                    </button>
-                    <button
-                        onclick={download_video}
-                        class="flex flex-1 transform items-center justify-center rounded-lg bg-gradient-to-r from-blue-600 to-blue-700 px-8 py-4 text-lg font-bold text-white shadow-lg transition-all duration-200 hover:-translate-y-1 hover:from-blue-700 hover:to-blue-800 hover:shadow-xl"
-                    >
-                        Download Video
-                    </button>
-                    <button
-                        onclick={() => "TODO: reset_output_wrapper"}
-                        class="flex flex-1 transform items-center justify-center rounded-lg bg-gradient-to-r from-gray-500 to-gray-600 px-8 py-4 text-lg font-bold text-white shadow-lg transition-all duration-200 hover:-translate-y-1 hover:from-gray-600 hover:to-gray-700 hover:shadow-xl"
-                    >
-                        Start Over
-                    </button>
-                </div>
-            </div>
-        {:else if temp_state.ffmpeg.video_file && temp_state.ffmpeg.srt_file && !temp_state.ffmpeg.is_processing}
-            <div class="rounded-lg border-2 border-dashed border-gray-300 bg-gray-50 py-12 text-center">
-                <h3 class="mb-2 text-lg font-semibold text-gray-700">Ready to Process!</h3>
-                <p class="text-gray-500">
-                    Configure your ASS subtitle styling above and click "Burn ASS Subtitles to Video" to create your
-                    styled subtitled video.
-                </p>
-            </div>
-        {:else}
-            <div class="rounded-lg border-2 border-dashed border-gray-300 bg-gray-50 py-12 text-center">
-                <h3 class="mb-2 text-lg font-semibold text-gray-700">Get Started</h3>
-                <p class="mb-4 text-gray-500">
-                    Upload your video file and SRT subtitle file to begin creating styled ASS subtitled videos.
-                </p>
-                <div class="flex justify-center space-x-4 text-sm text-gray-400">
-                    <span>Video</span>
-                    <span class="text-gray-300">→</span>
-                    <span>SRT → ASS</span>
-                    <span class="text-gray-300">→</span>
-                    <span>Output</span>
-                </div>
-            </div>
-        {/if}
+        <SubtitleOutput />
     </div>
 </div>
